Don't allow slashes in Path Templating segments
According to https://github.com/OAI/OpenAPI-Specification/issues/892
slashes aren't (yet?) supported in path templating parts, so change
the Path class to adhere to that.

This fixes a problem where declaring the paths
`/api/foo/{id}/` and `/api/foo/{id}/quux/` (in that order) would
cause the latter route to never be matched.